refactor: move polymorphic record retrival to tests
